fix(security): narrow getCurrentUser and harden Streamdown link handling
- Return explicit public user shape from auth.getCurrentUser (strip encryptedOpenRouterKey)
- Add returns validator for the query
- Share linkSafety config via streamdown-link-safety.tsx; use in reasoning and CoT
